// src/types/dogBreed.ts
export type DogBreed = {
  name: string;
  image_link: string;

  good_with_children: number;
  good_with_other_dogs: number;
  good_with_strangers: number;

  shedding: number;
  grooming: number;
  drooling: number;
  coat_length: number;

  playfulness: number;
  protectiveness: number;
  trainability: number;
  energy: number;
  barking: number;

  min_life_expectancy: number;
  max_life_expectancy: number;

  min_height_male: number;
  max_height_male: number;
  min_height_female: number;
  max_height_female: number;

  min_weight_male: number;
  max_weight_male: number;
  min_weight_female: number;
  max_weight_female: number;
};